requirements.txt
beautifulsoup4==4.13.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==43.0.0
dnspython==2.8.0
email_validator==2.2.0
filelock==3.20.0
idna==3.10
mail-parser==4.1.4
pycparser==2.23
pycryptodomex==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python-whois==0.9.3
requests==2.32.4
requests-file==3.0.1
six==1.17.0
soupsieve==2.8
tldextract==5.3.0
typing_extensions==4.15.0
urllib3==2.5.0
